+++ backends/arm/_passes/BUCK
@@ -24,6 +24,7 @@ fbcode_target(
         "//executorch/backends/arm:constants",
         "//executorch/backends/arm/tosa:utils",
         "//executorch/backends/arm/tosa/dialect:lib",
+        "//executorch/backends/transforms:permute_view_meta",
         "//executorch/backends/transforms:utils",
         "//executorch/exir:lib",
     ],
@@ -71,6 +72,7 @@ fbcode_target(
         "//executorch/backends/transforms:aten_to_dialect_pass",
         "//executorch/backends/arm/tosa:utils",
         "//executorch/backends/arm/tosa/dialect:lib",
+        "//executorch/backends/transforms:canonicalize_view_copy_permute_pass",
         "//executorch/backends/transforms:convert_conv1d_to_conv2d_pass",
         "//executorch/backends/transforms:fuse_view_copy",
         "//executorch/backends/transforms:remove_getitem_op",
